Extract commonalities from Phone, Email and Address classes into parent class called Contact
